requirements.txt
opencv-python>=4.5.0
ultralytics>=8.0.0
mysql-connector-python>=8.0.0
paho-mqtt>=1.6.0
numpy>=1.21.0
torch>=1.9.0
torchvision>=0.10.0
Pillow>=8.0.0
